+++ src/core/methods/initApp/initApp.ts
@@ -5,7 +5,10 @@ import { ToastManager } from 'core/managers/internal/ToastManager/ToastManager';
 import { login } from 'core/providers/DappProvider/helpers/login/login';
 import { restoreProvider } from 'core/providers/helpers/restoreProvider';
 import { ProviderFactory } from 'core/providers/ProviderFactory';
-import { ProviderTypeEnum } from 'core/providers/types/providerFactory.types';
+import {
+  ICustomProvider,
+  ProviderTypeEnum
+} from 'core/providers/types/providerFactory.types';
 import { getDefaultNativeAuthConfig } from 'services/nativeAuth/methods/getDefaultNativeAuthConfig';
 import { NativeAuthConfigType } from 'services/nativeAuth/nativeAuth.types';
 import { initializeNetwork } from 'store/actions';
@@ -111,12 +114,17 @@ export async function initApp({
     signTransactionsStateManager.init()
   ]);
 
-  const usedProviders = [
+  const usedProviders: ICustomProvider[] = [
     ...((safeWindow as any)?.multiversx?.providers ?? []),
     ...(customProviders || [])
   ];
 
-  ProviderFactory.customProviders(usedProviders || []);
+  const uniqueProviders = usedProviders.filter(
+    (provider, index, arr) =>
+      index === arr.findIndex((item) => item.type === provider.type)
+  );
+
+  ProviderFactory.customProviders = uniqueProviders || [];
 
   if (isLoggedIn && !isAppInitialized) {
     await restoreProvider();